Added Frontend UI

+++ backend/controllers/goalController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goal");
+const User = require("../models/user");
 
 module.exports.getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.json(goals);
 });
 
@@ -13,6 +14,7 @@ module.exports.createGoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
@@ -23,6 +25,19 @@ module.exports.updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400).json({ message: "Goal not found" });
   }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User Not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User Not Authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -32,8 +47,21 @@ module.exports.updateGoal = asyncHandler(async (req, res) => {
 module.exports.deleteGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
   if (!goal) {
-    res.status(400).json({ message: "Goal not found" });
+    return res.status(400).json({ message: "Goal not found" });
   }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User Not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User Not Authorized");
+  }
+
   const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
-  res.json(deletedGoal);
+  res.json({ id: req.params.id });
 });
